requirements.txt
asgiref==3.8.1
certifi==2025.4.26
cffi==1.17.1
chardet==3.0.4
charset-normalizer==3.4.2
click==8.2.0
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
cryptography==44.0.3
defusedxml==0.7.1
dj-database-url==0.5.0
Django==3.2
django-cors-headers==3.7.0
django-templated-mail==1.1.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
djoser==2.1.0
exceptiongroup==1.2.2
googletrans==4.0.0rc1
gunicorn==23.0.0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2025.1.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
itypes==1.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
oauthlib==3.2.2
packaging==25.0
psycopg2==2.9.10
pycparser==2.22
PyJWT==2.10.1
python-dotenv==1.0.1
python3-openid==3.2.0
pytz==2025.2
requests==2.32.3
requests-oauthlib==2.0.0
rfc3986==1.5.0
six==1.17.0
sniffio==1.3.1
social-auth-app-django==4.0.0
social-auth-core==4.6.1
sqlparse==0.5.3
typing_extensions==4.13.1
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.4.0
uvicorn==0.34.2
whitenoise==6.9.0
